requirements.txt
azure-functions>=1.18.0
azure-ai-documentintelligence
pyodbc
azure-storage-blob
azure-identity
openai